added signals and started to use t_minishell

+++ src/euthanasia/euthanasia.c
@@ -6,13 +6,11 @@
 /*   By: geudes <geudes@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/21 02:34:49 by geudes            #+#    #+#             */
-/*   Updated: 2023/06/25 07:57:46 by geudes           ###   ########.fr       */
+/*   Updated: 2023/06/29 09:11:28 by geudes           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <signal.h>
-#include <stdlib.h>
-#include <unistd.h>
+#include "minishell.h"
 
 struct	s_mytimespec
 {
@@ -20,7 +18,7 @@ struct	s_mytimespec
 	long int			nano;
 };
 
-void	myusleep(struct mytimespec *var)
+void	myusleep(struct s_mytimespec *var)
 {
 	asm volatile ("mov $35, %rax;"
 		"mov -8(%rsp), %rdi;"
@@ -32,12 +30,12 @@ void	myusleep(struct mytimespec *var)
 void	handle(int sig, siginfo_t *info, void *context)
 {
 	(void)context;
-	(void)sig;
-	kill(info->si_pid, SIGKILL);
+	if (sig == SIGUSR1)
+		kill(info->si_pid, SIGKILL);
 	exit(0);
 }
 
-__attribute__((constructor))
+// __attribute__((constructor))
 void	sig_eutanasia(void)
 {
 	struct sigaction	sigusr;
@@ -46,6 +44,10 @@ void	sig_eutanasia(void)
 	sigusr.sa_sigaction = handle;
 	sigemptyset(&sigusr.sa_mask);
 	if (sigaction(SIGUSR1, &sigusr, 0))
+		exit(0);
+	if (sigaction(SIGQUIT, &sigusr, 0))
+		exit(0);
+	if (sigaction(SIGINT, &sigusr, 0))
 		exit(0);
 }
 
@@ -64,4 +66,5 @@ int	euthanasia(void)
 	one_sec.nano = 0;
 	while (1)
 		myusleep(&one_sec);
+	return (0);
 }
